requirements.txt
boto3==1.34.127
botocore==1.34.127
moviepy==1.0.3
playwright==1.44.0
praw==7.7.1
prawcore~=2.3.0
requests==2.32.3
toml==0.10.2
Pillow==9.4.0
Markdown==3.4.1
tinydb==4.7.0
imageio==2.24.0
imageio-ffmpeg==0.4.8
numpy==1.26.4
mutagen==1.46.0
beautifulsoup4==4.11.1
pydub==0.25.1
clean-text==0.6.0
unidecode==1.3.8
spacy==3.7.5
openai-whisper==20231117
python-dotenv==1.0.1